Moved CubicEoS to another file and added properties. NEEDSFIX: Check the parameters input of the different new properties like pressure
